Finish cli stuff
