--- notebooks/.env.default
@@ -1,1 +1,0 @@
-GIRDER_API_KEY=CHANGEME
