+++ server/dive_server/viame.py
@@ -5,6 +5,7 @@ from girder.api.describe import Description, autoDescribeRoute, describeRoute
 from girder.api.rest import Resource
 from girder.constants import AccessType
 from girder.exceptions import RestException
+from girder.models.file import File
 from girder.models.folder import Folder
 from girder.models.item import Item
 from girder.models.setting import Setting
@@ -39,10 +40,10 @@ from dive_utils.constants import (
     videoRegex,
     ymlRegex,
 )
+from dive_utils.serializers import meva
 from dive_utils.types import AvailableJobSchema, PipelineDescription
 
 from .pipelines import load_pipelines, run_pipeline
-from .serializers import meva as meva_serializer
 from .training import ensure_csv_detections_file, training_output_folder
 from .transforms import GetPathFromItemId
 from .utils import (
@@ -430,8 +431,12 @@ class Viame(Resource):
             ymlItems = Folder().childItems(folder, filters={"lowerName": {"$regex": ymlRegex}})
             if ymlItems.count() > 0:
                 # There might be up to 3 yamls
-                allFiles = [Item().childFiles(item)[0] for item in ymlItems]
-                saveTracks(folder, meva_serializer.load_kpf_as_tracks(allFiles), user)
+                def make_file_generator(item):
+                    file = Item().childFiles(item)[0]
+                    return File().download(file, headers=False)()
+
+                allFiles = [make_file_generator(item) for item in ymlItems]
+                saveTracks(folder, meva.load_kpf_as_tracks(allFiles), user)
                 ymlItems.rewind()
                 for item in ymlItems:
                     Item().move(item, auxiliary)
