+++ server/setup.py
@@ -20,20 +20,12 @@ requirements = [
     "urllib3<1.26",
 ]
 
-# For running tests, linting, etc.
 dev_requirements = [
+    "numpy",
+    "opencv-python",
     "pytest",
     "pytest-cov",
     "tox",
-]
-
-# For running debugging command line tools
-# and notebooks
-debug_requirements = [
-    "ipykernel",
-    "opencv-python",
-    "numpy",
-    "python-dotenv",
 ]
 
 setup(
@@ -59,7 +51,7 @@ setup(
     include_package_data=True,
     zip_safe=False,
     entry_points={
-        "console_scripts": ["dive=scripts.cli:cli"],
+        "console_scripts": ["dive=scripts.cli_dev:cli"],
         "girder.plugin": [
             "dive_server = dive_server:GirderPlugin",
             "bucket_notifications = bucket_notifications:GirderPlugin",
@@ -68,8 +60,5 @@ setup(
         "girder_worker_plugins": ["dive_tasks = dive_tasks:DIVEPlugin"],
     },
     install_requires=requirements,
-    extras_require={
-        "dev": dev_requirements,
-        "debug": debug_requirements,
-    },
+    extras_require={"dev": dev_requirements},
 )
