+++ server/tox.ini
@@ -54,6 +54,14 @@ deps =
 commands =
     mkdocs serve {posargs}
 
+[testenv:buildcli]
+extra =
+    dev
+deps =
+    pyinstaller
+commands =
+    pyinstaller --onefile scripts/cli.py
+
 [flake8]
 max-line-length = 100
 show-source = True
